refactor(auth): align validator namespaces with command structure
- Move Login, Register, and GoogleLogin validators to their command namespaces
