+++ src/FinanceTracker.Application/Features/Auth/Commands/GoogleLogin/GoogleLoginCommandValidator.cs
@@ -1,7 +1,6 @@
-﻿using FinanceTracker.Application.Features.Auth.Commands.GoogleLogin;
-using FluentValidation;
+﻿using FluentValidation;
 
-namespace FinanceTracker.Application.Features.Auth.Validators;
+namespace FinanceTracker.Application.Features.Auth.Commands.GoogleLogin;
 
 public class GoogleLoginCommandValidator : AbstractValidator<GoogleLoginCommand>
 {
